kala/requirements.txt
amqp
celery
defusedxml
distribute
django-celery
django-countries
django-timezone-field
django-localflavor
django-localflavor-us
django-uuidfield
djangorestframework
ipython
psycopg2
python-dateutil
pytz
requests
six
task
